// frontend/src/types/leave.ts
export interface PublicHoliday {
  id: string;
  name: string;
  date: string;
  isRecurring: boolean;
  branch: string | null;
}

export interface SavePublicHolidayRequest {
  name: string;
  date: string;
  isRecurring: boolean;
  branch?: string | null;
}